United States Baby Name data

Returns detailed data about the popularity of various baby names in the United States.

us_baby_names(type = "national", include_features = T)

Arguments

type
the type of data options include

`national` - National popularity by state, sex and year - default

`state` - Popularity by state, sex and year

`territory` - Popularity by territory, sex and year

include_features
if `TRUE` includes features about the popularity of the name by sex

Value

a `tibble` or a nested `tibble` if type length exceeds 1

Details

Includes access to national, state, and territorial data.

Examples

library(babynamer)

tbl_usa <- us_baby_names(type = "national")
tbl_state <- us_baby_names(type = "state")
tbl_territory <- us_baby_names(type = "territory")
us_baby_names <- us_baby_names(type = c("national","state", "territory"))